src/coffee-shop/pyproject.toml
[tool.poetry]
name = "coffee-shop"
version = "0.1.0"
description = "Coffee shop API"
authors = ["Guilherme Arthur de Carvalho <gui.ifsp11@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.0"
flask = "^1.0"
flask-sqlalchemy = "^2.3"
psycopg2 = "^2.7"
flask-jwt-extended = "^3.13"
flask-migrate = "^2.2"
flask-restful = "^0.3.6"
flask-marshmallow = "^0.9.0"
marshmallow-sqlalchemy = "^0.14.1"
marshmallow = "^2.16"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3.5"
flask-shell-ipython = "^0.3.1"
ipdb = "^0.11.0"
pytest-flask = "^0.13.0"
